created async transfers call.
	modified:   gpu/common.cu
	modified:   gpu/common.cuh
	modified:   gpu/gpu_struct.cu
	modified:   gpu/gpu_struct.cuh
	modified:   gpu/kernel_init_edges.cu
	modified:   gpu/kernel_multi_search.cu
	modified:   include/bit_vector.h
	modified:   mcb/mcb_cycle_deg_two_remove.cpp

+++ gpu/common.cuh
@@ -23,7 +23,7 @@ extern "C" unsigned* allocate_pinned_memory(int chunk, int nodes);
 
 extern "C" void free_pinned_memory(unsigned *);
 
-extern "C" int calculate_chunk_size(int num_nodes, int num_edges,
-		int size_vector, int nstream);
+extern "C" size_t calculate_chunk_size(size_t num_nodes, size_t num_edges,
+		size_t size_vector, size_t nstream);
 
 #endif
